Rename common to lib #1
